Added donation link
It just struck me that Adblock Plus has a "Known Adblock Plus subscriptions" archive, meaning that I could've submitted this list to that place too. Since that archive doesn't appear to me to make use of dedicated donation links, I hope it wouldn't be frowned upon if I added said link to the filter list itself instead.

Also changed the version number to the European date format, and added a Norwegian note that explains that people can also take contact with me in the Scandinavian languages.