Update ExperimentalNorwegianList.txt
So the clock is 07:30, and after having pulled an all-nighter, what I decided to do on a morning like this is to... look through all the other filter lists out there that are about everything from cookie notices to malware, in order to borrow Norway-relevant filters from them. I really hope they work as well as they should.