Narrowed down the cashback whitelist
To ease any privacy concerns out there a bit, I've attempted to make the whitelist entries more narrow so that they're more likely to only apply to Norwegian shopping sites.

Also, a heads up: Startshop.no appears to be a dead prospect that no longer pays out cashback, so I advice against using that site anymore. I was even tempted to add a blocking rule for their entire site, but chose not to do so for now.