Update Twitter De-Politificator.txt
How is it even possible for someone to be a brony and a white supremacist at the same time? How can they just rotate every 2min between posting fluff fanart of FiM, then posting racist stuff so bad it can't even be recited on most sites, and then back again? Did they even watch 5min of the gods-damn show they claim to be fans of?